Moved WordSpec tests classes to wordspec-test under jvm.
