+++ build.sbt
@@ -22,6 +22,7 @@ lazy val scalatestFreeSpecTest = ScalatestBuild.scalatestFreeSpecTest
 lazy val scalatestFunSpecTest = ScalatestBuild.scalatestFunSpecTest
 lazy val scalatestFunSuiteTest = ScalatestBuild.scalatestFunSuiteTest
 lazy val scalatestPropSpecTest = ScalatestBuild.scalatestPropSpecTest
+lazy val scalatestWordSpecTest = ScalatestBuild.scalatestWordSpecTest
 lazy val scalatestJS = ScalatestBuild.scalatestJS
 lazy val scalatestNative = ScalatestBuild.scalatestNative
 lazy val scalatestDotty = ScalatestBuild.scalatestDotty
