+++ jvm/wordspec-test/src/test/scala/org/scalatest/wordspec/AsyncWordSpecLikeSpec2.scala
@@ -13,59 +13,55 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.wordspec
 
 import org.scalatest._
-import SharedHelpers.EventRecordingReporter
+import org.scalatest.SharedHelpers.EventRecordingReporter
 import scala.concurrent.{ExecutionContext, Promise, Future}
 import org.scalatest.concurrent.SleepHelper
 import org.scalatest.events.{InfoProvided, MarkupProvided}
 
 import scala.util.Success
-import org.scalatest
-import org.scalatest.wordspec
+import org.scalatest.funspec.AsyncFunSpec
+import org.scalatest.wordspec.AsyncWordSpecLike
 
-class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
+class AsyncWordSpecLikeSpec2 extends AsyncFunSpec {
 
   describe("AsyncWordSpecLike") {
 
     it("can be used for tests that return Future under parallel async test execution") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike with ParallelTestExecution {
-
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class ExampleSpec extends AsyncWordSpecLike with ParallelTestExecution {
 
         val a = 1
 
-        "test 1" in { fixture =>
+        "test 1" in {
           Future {
             assert(a == 1)
           }
         }
 
-        "test 2" in { fixture =>
+        "test 2" in {
           Future {
             assert(a == 2)
           }
         }
 
-        "test 3" in { fixture =>
+        "test 3" in {
           Future {
             pending
           }
         }
 
-        "test 4" in { fixture =>
+        "test 4" in {
           Future {
-            cancel
+            cancel()
           }
         }
 
-        "test 5" ignore { fixture =>
+        "test 5" ignore {
           Future {
-            cancel
+            cancel()
           }
         }
 
@@ -94,32 +90,28 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike with ParallelTestExecution {
-
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class ExampleSpec extends AsyncWordSpecLike with ParallelTestExecution {
 
         val a = 1
 
-        "test 1" in { fixture =>
+        "test 1" in {
           assert(a == 1)
         }
 
-        "test 2" in { fixture =>
+        "test 2" in {
           assert(a == 2)
         }
 
-        "test 3" in { fixture =>
+        "test 3" in {
           pending
         }
 
-        "test 4" in { fixture =>
-          cancel
+        "test 4" in {
+          cancel()
         }
 
-        "test 5" ignore { fixture =>
-          cancel
+        "test 5" ignore {
+          cancel()
         }
 
         override def newInstance = new ExampleSpec
@@ -149,13 +141,9 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
+      class ExampleSpec extends AsyncWordSpecLike {
 
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
-
-        "test 1" in { fixture =>
+        "test 1" in {
           Future {
             SleepHelper.sleep(30)
             assert(count == 0)
@@ -164,7 +152,7 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
           }
         }
 
-        "test 2" in { fixture =>
+        "test 2" in {
           Future {
             assert(count == 1)
             SleepHelper.sleep(50)
@@ -173,7 +161,7 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
           }
         }
 
-        "test 3" in { fixture =>
+        "test 3" in {
           Future {
             assert(count == 2)
           }
@@ -196,27 +184,23 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
+      class ExampleSpec extends AsyncWordSpecLike {
 
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
-
-        "test 1" in { fixture =>
+        "test 1" in {
           SleepHelper.sleep(30)
           assert(count == 0)
           count = 1
           Succeeded
         }
 
-        "test 2" in { fixture =>
+        "test 2" in {
           assert(count == 1)
           SleepHelper.sleep(50)
           count = 2
           Succeeded
         }
 
-        "test 3" in { fixture =>
+        "test 3" in {
           assert(count == 2)
         }
 
@@ -241,20 +225,16 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
       var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
+      class ExampleSpec extends AsyncWordSpecLike {
 
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
-
-        "test 1" in { fixture =>
+        "test 1" in {
           Future {
             test1Thread = Some(Thread.currentThread)
             succeed
           }
         }
 
-        "test 2" in { fixture =>
+        "test 2" in {
           Future {
             test2Thread = Some(Thread.currentThread)
             succeed
@@ -288,13 +268,9 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
       @volatile var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
+      class ExampleSpec extends AsyncWordSpecLike {
 
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
-
-        "test 1" in { fixture =>
+        "test 1" in {
           val promise = Promise[Assertion]
           val timer = new java.util.Timer
           timer.schedule(
@@ -311,7 +287,7 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
           }
         }
 
-        "test 2" in { fixture =>
+        "test 2" in {
           val promise = Promise[Assertion]
           val timer = new java.util.Timer
           timer.schedule(
@@ -351,15 +327,11 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
+      class ExampleSpec extends AsyncWordSpecLike {
 
         // Note we get a StackOverflowError with the following execution
         // context.
         // override implicit def executionContext: ExecutionContext = new ExecutionContext { def execute(runnable: Runnable) = runnable.run; def reportFailure(cause: Throwable) = () }
-
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
 
         def sum(xs: List[Int]): Future[Int] =
           xs match {
@@ -367,7 +339,7 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
             case x :: xs => Future(x).flatMap(xx => sum(xs).map(xxx => xx + xxx))
           }
 
-        "test 1" in { fixture =>
+        "test 1" in {
           val fut: Future[Int] = sum((1 to 50000).toList)
           fut.map(total => assert(total == 1250025000))
         }
@@ -387,27 +359,23 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should run tests that returns Future and report their result in serial") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
+      class ExampleSpec extends AsyncWordSpecLike {
 
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
-
-        "test 1" in { fixture =>
+        "test 1" in {
           Future {
             SleepHelper.sleep(60)
             succeed
           }
         }
 
-        "test 2" in { fixture =>
+        "test 2" in {
           Future {
             SleepHelper.sleep(30)
             succeed
           }
         }
 
-        "test 3" in { fixture =>
+        "test 3" in {
           Future {
             succeed
           }
@@ -418,6 +386,9 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
+      // SKIP-SCALATESTJS,NATIVE-START
+      status.waitUntilCompleted()
+      // SKIP-SCALATESTJS,NATIVE-END
 
       val promise = Promise[EventRecordingReporter]
       status whenCompleted { _ => promise.success(rep) }
@@ -435,23 +406,19 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should run tests that does not return Future and report their result in serial") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
+      class ExampleSpec extends AsyncWordSpecLike {
 
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
-
-        "test 1" in { fixture =>
+        "test 1" in {
           SleepHelper.sleep(60)
           succeed
         }
 
-        "test 2" in { fixture =>
+        "test 2" in {
           SleepHelper.sleep(30)
           succeed
         }
 
-        "test 3" in { fixture =>
+        "test 3" in {
           succeed
         }
 
@@ -460,6 +427,9 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
+      // SKIP-SCALATESTJS,NATIVE-START
+      status.waitUntilCompleted()
+      // SKIP-SCALATESTJS,NATIVE-END
 
       val promise = Promise[EventRecordingReporter]
       status whenCompleted { _ => promise.success(rep) }
@@ -476,10 +446,7 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
         info(
           "hi there"
         )
@@ -499,16 +466,14 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in scope body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
           info(
             "hi there"
           )
 
-          "test 1" in { fixture => succeed }
+          "test 1" in { succeed }
         }
       }
       val suite = new MySuite
@@ -526,12 +491,10 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
-          "test 1" in { fixture =>
+          "test 1" in {
             info("hi there")
             succeed
           }
@@ -558,12 +521,10 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in Future returned by test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
-          "test 1" in { fixture =>
+          "test 1" in {
             Future {
               info("hi there")
               succeed
@@ -592,10 +553,7 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
         note(
           "hi there"
         )
@@ -615,16 +573,14 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in scope body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
           note(
             "hi there"
           )
 
-          "test 1" in { fixture => succeed }
+          "test 1" in { succeed }
         }
       }
       val suite = new MySuite
@@ -642,12 +598,10 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
-          "test 1" in { fixture =>
+          "test 1" in {
             note("hi there")
             succeed
           }
@@ -667,12 +621,10 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in Future returned by test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
-          "test 1" in { fixture =>
+          "test 1" in {
             Future {
               note("hi there")
               succeed
@@ -694,10 +646,7 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
         alert(
           "hi there"
         )
@@ -717,16 +666,14 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in scope body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
           alert(
             "hi there"
           )
 
-          "test 1" in { fixture => succeed }
+          "test 1" in { succeed }
         }
       }
       val suite = new MySuite
@@ -744,12 +691,10 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
-          "test 1" in { fixture =>
+          "test 1" in {
             alert("hi there")
             succeed
           }
@@ -769,12 +714,10 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in Future returned by test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
-          "test 1" in { fixture =>
+          "test 1" in {
             Future {
               alert("hi there")
               succeed
@@ -796,10 +739,7 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
         markup(
           "hi there"
         )
@@ -819,16 +759,14 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in scope body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
           markup(
             "hi there"
           )
 
-          "test 1" in { fixture => succeed }
+          "test 1" in { succeed }
         }
       }
       val suite = new MySuite
@@ -846,12 +784,10 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
-          "test 1" in { fixture =>
+          "test 1" in {
             markup("hi there")
             succeed
           }
@@ -878,12 +814,10 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome =
-          test("testing")
+      class MySuite extends AsyncWordSpecLike  {
+
         "test feature" should {
-          "test 1" in { fixture =>
+          "test 1" in {
             Future {
               markup("hi there")
               succeed
@@ -912,28 +846,24 @@ class AsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should allow other execution context to be used") {
-      class TestSpec extends wordspec.FixtureAsyncWordSpecLike {
+      class TestSpec extends AsyncWordSpecLike {
         // SKIP-SCALATESTJS,NATIVE-START
         override implicit val executionContext = scala.concurrent.ExecutionContext.Implicits.global
         // SKIP-SCALATESTJS,NATIVE-END
         // SCALATESTJS-ONLY override implicit val executionContext = scala.scalajs.concurrent.JSExecutionContext.runNow
-
-        type FixtureParam = String
-        def withFixture(test: OneArgAsyncTest): FutureOutcome = { test("hi") }
-
         val a = 1
         "feature 1" should {
-          "test A" in { fixture =>
+          "test A" in {
             Future { assert(a == 1) }
           }
         }
         "feature 2" should {
-          "test B" in { fixture =>
+          "test B" in {
             Future { assert(a == 1) }
           }
         }
         "feature 3" should {
-          "test C" in { fixture =>
+          "test C" in {
             Future { assert(a == 1) }
           }
         }
