+++ jvm/wordspec-test/src/test/scala/org/scalatest/wordspec/FixtureAsyncWordSpecLikeSpec2.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.wordspec
 
 import org.scalatest._
 import SharedHelpers.EventRecordingReporter
@@ -25,13 +25,13 @@ import scala.util.Success
 import org.scalatest
 import org.scalatest.wordspec
 
-class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
+class FixtureAsyncWordSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
-  describe("AsyncWordSpec") {
+  describe("AsyncWordSpecLike") {
 
     it("can be used for tests that return Future under parallel async test execution") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec with ParallelTestExecution {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -94,7 +94,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec with ParallelTestExecution {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -149,7 +149,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -196,7 +196,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -241,7 +241,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
       var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -288,7 +288,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
       @volatile var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -351,7 +351,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
 
         // Note we get a StackOverflowError with the following execution
         // context.
@@ -387,7 +387,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should run tests that returns Future and report their result in serial") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -435,7 +435,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should run tests that does not return Future and report their result in serial") {
 
-      class ExampleSpec extends wordspec.FixtureAsyncWordSpec {
+      class ExampleSpec extends wordspec.FixtureAsyncWordSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -476,7 +476,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -499,7 +499,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in scope body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -526,7 +526,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -558,7 +558,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in Future returned by test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -592,7 +592,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -615,7 +615,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in scope body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -642,7 +642,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -667,7 +667,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in Future returned by test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -694,7 +694,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -717,7 +717,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in scope body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -744,7 +744,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -769,7 +769,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in Future returned by test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -796,7 +796,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -819,7 +819,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in scope body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -846,7 +846,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -878,7 +878,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by test body") {
-      class MySuite extends wordspec.FixtureAsyncWordSpec  {
+      class MySuite extends wordspec.FixtureAsyncWordSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -912,7 +912,7 @@ class AsyncWordSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should allow other execution context to be used") {
-      class TestSpec extends wordspec.FixtureAsyncWordSpec {
+      class TestSpec extends wordspec.FixtureAsyncWordSpecLike {
         // SKIP-SCALATESTJS,NATIVE-START
         override implicit val executionContext = scala.concurrent.ExecutionContext.Implicits.global
         // SKIP-SCALATESTJS,NATIVE-END
