+++ jvm/wordspec-test/src/test/scala/org/scalatest/wordspec/WordSpecImportedMatchersSpec.scala
@@ -13,11 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.wordspec
 
-import org.scalatest.matchers.should.Matchers
+import org.scalatest.matchers.should.Matchers._
 import org.scalatest.wordspec.AnyWordSpec
-class WordSpecMixedInMatchersSpec extends AnyWordSpec with Matchers {
+
+/*
+This tests that matchers works with WordSpec when matchers are imported,
+something that broke in 2.1.RC1.
+*/
+class WordSpecImportedMatchersSpec extends AnyWordSpec {
   "This spec" should {
     "work OK" in {
       "hello" should startWith ("he")
