+++ jvm/wordspec-test/src/test/scala/org/scalatest/wordspec/WordSpecSpec.scala
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.wordspec
 
 // elements
-import SharedHelpers._
+import org.scalatest._
+import org.scalatest.SharedHelpers._
 import org.scalatest.events._
 import org.scalactic.Prettifier
 import java.awt.AWTError
